Replace absolute paths with relative paths.
